requirements.txt
bleak==2.1.1
matplotlib==3.10.8
numpy==2.3.1
opencv-python==4.13.0.92
pillow
rich
scipy==1.17.1
